Moved get_free_port function to utils directory to make it available other OpenMDAO needs and also OpenMDAO related projects like Aviary
